+++ packages/bluesky/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - Added `labelerDid` on `ContentLabelPreference`. ([#1316](https://github.com/myConsciousness/atproto.dart/issues/1316))
 - Added `.associated` property on `ActorProfile`. ([#1316](https://github.com/myConsciousness/atproto.dart/issues/1316))
+- Supported Labeler Service as `LabelerService`. ([#1316](https://github.com/myConsciousness/atproto.dart/issues/1316))
+  - Added `.labeler` on `Bluesky` object
+    - `.getServices`
+    - `.service`
+  - Added `.isLabelerService` and `.isNotLabelerService` on AT URI extension
+  - Added `LabelerServiceRecord` object
+  - Added `.onCreateLabelerService` and `.onDeleteLabelerService` on `RepoCommitAdaptor` object
 
 ## v0.15.12